LayoutTests/platform/mac/accessibility/image-map1-expected.txt

 
This tests that you can reach the links within an image map.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.childAtIndex(0).role is 'AXRole: AXLink'
PASS body.childAtIndex(0).title is 'AXTitle: Link1'
PASS body.childAtIndex(1).role is 'AXRole: AXLink'
PASS body.childAtIndex(1).title is 'AXTitle: Link2'
PASS body.childAtIndex(2).role is 'AXRole: AXLink'
PASS body.childAtIndex(2).title is 'AXTitle: Link3'
PASS successfullyParsed is true

TEST COMPLETE
